e2e: Align uiautomator package with module layout and document the E2E split
- Rename io.getstream.chat.android.compose.uiautomator to io.getstream.chat.android.e2e.test.uiautomator so the package matches its location and the module's other packages (mockserver, robots, rules)
- Move VisualOrder into the e2e-test module's uiautomator toolkit (now public), since it is app-agnostic infrastructure
- Document the module-vs-sample E2E convention in AGENTS.md
