+++ stream-chat-android-compose-sample/src/androidTestE2eDebug/kotlin/io/getstream/chat/android/compose/robots/UserRobotMessageListAsserts.kt
@@ -23,24 +23,24 @@ import io.getstream.chat.android.compose.pages.MessageListPage
 import io.getstream.chat.android.compose.pages.MessageListPage.Composer
 import io.getstream.chat.android.compose.pages.MessageListPage.MessageList.Message
 import io.getstream.chat.android.compose.pages.ThreadPage
-import io.getstream.chat.android.compose.uiautomator.appContext
-import io.getstream.chat.android.compose.uiautomator.device
-import io.getstream.chat.android.compose.uiautomator.findObject
-import io.getstream.chat.android.compose.uiautomator.findObjects
-import io.getstream.chat.android.compose.uiautomator.height
-import io.getstream.chat.android.compose.uiautomator.isDisplayed
-import io.getstream.chat.android.compose.uiautomator.isEnabled
-import io.getstream.chat.android.compose.uiautomator.retryOnStaleObjectException
-import io.getstream.chat.android.compose.uiautomator.seconds
-import io.getstream.chat.android.compose.uiautomator.wait
-import io.getstream.chat.android.compose.uiautomator.waitForCount
-import io.getstream.chat.android.compose.uiautomator.waitForText
-import io.getstream.chat.android.compose.uiautomator.waitToAppear
-import io.getstream.chat.android.compose.uiautomator.waitToAppearBottomUp
-import io.getstream.chat.android.compose.uiautomator.waitToDisappear
 import io.getstream.chat.android.e2e.test.mockserver.MessageDeliveryStatus
 import io.getstream.chat.android.e2e.test.mockserver.ReactionType
 import io.getstream.chat.android.e2e.test.robots.ParticipantRobot
+import io.getstream.chat.android.e2e.test.uiautomator.appContext
+import io.getstream.chat.android.e2e.test.uiautomator.device
+import io.getstream.chat.android.e2e.test.uiautomator.findObject
+import io.getstream.chat.android.e2e.test.uiautomator.findObjects
+import io.getstream.chat.android.e2e.test.uiautomator.height
+import io.getstream.chat.android.e2e.test.uiautomator.isDisplayed
+import io.getstream.chat.android.e2e.test.uiautomator.isEnabled
+import io.getstream.chat.android.e2e.test.uiautomator.retryOnStaleObjectException
+import io.getstream.chat.android.e2e.test.uiautomator.seconds
+import io.getstream.chat.android.e2e.test.uiautomator.wait
+import io.getstream.chat.android.e2e.test.uiautomator.waitForCount
+import io.getstream.chat.android.e2e.test.uiautomator.waitForText
+import io.getstream.chat.android.e2e.test.uiautomator.waitToAppear
+import io.getstream.chat.android.e2e.test.uiautomator.waitToAppearBottomUp
+import io.getstream.chat.android.e2e.test.uiautomator.waitToDisappear
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotEquals
