+++ stream-chat-android-compose-sample/src/androidTestE2eDebug/kotlin/io/getstream/chat/android/compose/tests/AuthTests.kt
@@ -19,14 +19,14 @@ package io.getstream.chat.android.compose.tests
 import io.getstream.chat.android.compose.pages.JwtPage
 import io.getstream.chat.android.compose.robots.UserRobot
 import io.getstream.chat.android.compose.sample.ui.InitTestActivity
-import io.getstream.chat.android.compose.uiautomator.device
-import io.getstream.chat.android.compose.uiautomator.disableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.enableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.goToBackground
-import io.getstream.chat.android.compose.uiautomator.goToForeground
-import io.getstream.chat.android.compose.uiautomator.isDisplayed
-import io.getstream.chat.android.compose.uiautomator.seconds
-import io.getstream.chat.android.compose.uiautomator.waitToAppear
+import io.getstream.chat.android.e2e.test.uiautomator.device
+import io.getstream.chat.android.e2e.test.uiautomator.disableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.enableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.goToBackground
+import io.getstream.chat.android.e2e.test.uiautomator.goToForeground
+import io.getstream.chat.android.e2e.test.uiautomator.isDisplayed
+import io.getstream.chat.android.e2e.test.uiautomator.seconds
+import io.getstream.chat.android.e2e.test.uiautomator.waitToAppear
 import io.qameta.allure.kotlin.Allure.step
 import io.qameta.allure.kotlin.AllureId
 import org.junit.Assert.assertTrue
