+++ stream-chat-android-compose-sample/src/androidTestE2eDebug/kotlin/io/getstream/chat/android/compose/tests/ChannelListTests.kt
@@ -21,11 +21,11 @@ import io.getstream.chat.android.compose.robots.assertMessageDeliveryStatus
 import io.getstream.chat.android.compose.robots.assertMessageInChannelPreview
 import io.getstream.chat.android.compose.robots.assertMessagePreviewTimestamp
 import io.getstream.chat.android.compose.sample.ui.InitTestActivity
-import io.getstream.chat.android.compose.uiautomator.device
-import io.getstream.chat.android.compose.uiautomator.disableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.enableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.seconds
 import io.getstream.chat.android.e2e.test.mockserver.MessageDeliveryStatus
+import io.getstream.chat.android.e2e.test.uiautomator.device
+import io.getstream.chat.android.e2e.test.uiautomator.disableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.enableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.seconds
 import io.qameta.allure.kotlin.Allure.step
 import io.qameta.allure.kotlin.AllureId
 import org.junit.Ignore
