+++ stream-chat-android-compose-sample/src/androidTestE2eDebug/kotlin/io/getstream/chat/android/compose/tests/MessageListTests.kt
@@ -35,13 +35,13 @@ import io.getstream.chat.android.compose.robots.assertThreadMessage
 import io.getstream.chat.android.compose.robots.assertThreadReplyLabelOnParentMessage
 import io.getstream.chat.android.compose.robots.assertTypingIndicator
 import io.getstream.chat.android.compose.sample.ui.InitTestActivity
-import io.getstream.chat.android.compose.uiautomator.device
-import io.getstream.chat.android.compose.uiautomator.disableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.enableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.goToBackground
-import io.getstream.chat.android.compose.uiautomator.goToForeground
 import io.getstream.chat.android.e2e.test.mockserver.MessageDeliveryStatus
 import io.getstream.chat.android.e2e.test.mockserver.forbiddenWord
+import io.getstream.chat.android.e2e.test.uiautomator.device
+import io.getstream.chat.android.e2e.test.uiautomator.disableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.enableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.goToBackground
+import io.getstream.chat.android.e2e.test.uiautomator.goToForeground
 import io.qameta.allure.kotlin.Allure.step
 import io.qameta.allure.kotlin.AllureId
 import org.junit.Ignore
