+++ stream-chat-android-compose-sample/src/androidTestE2eDebug/kotlin/io/getstream/chat/android/compose/tests/ReactionsTests.kt
@@ -19,12 +19,12 @@ package io.getstream.chat.android.compose.tests
 import io.getstream.chat.android.compose.pages.MessageListPage.MessageList.Message
 import io.getstream.chat.android.compose.robots.assertReaction
 import io.getstream.chat.android.compose.sample.ui.InitTestActivity
-import io.getstream.chat.android.compose.uiautomator.device
-import io.getstream.chat.android.compose.uiautomator.disableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.enableInternetConnection
-import io.getstream.chat.android.compose.uiautomator.seconds
-import io.getstream.chat.android.compose.uiautomator.wait
 import io.getstream.chat.android.e2e.test.mockserver.ReactionType
+import io.getstream.chat.android.e2e.test.uiautomator.device
+import io.getstream.chat.android.e2e.test.uiautomator.disableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.enableInternetConnection
+import io.getstream.chat.android.e2e.test.uiautomator.seconds
+import io.getstream.chat.android.e2e.test.uiautomator.wait
 import io.qameta.allure.kotlin.Allure.step
 import io.qameta.allure.kotlin.AllureId
 import org.junit.Test
