+++ stream-chat-android-compose-sample/src/androidTestE2eDebug/kotlin/io/getstream/chat/android/compose/tests/StreamTestCase.kt
@@ -25,14 +25,14 @@ import android.annotation.SuppressLint
 import android.content.Intent
 import io.getstream.chat.android.compose.robots.UserRobot
 import io.getstream.chat.android.compose.sample.ui.InitTestActivity
-import io.getstream.chat.android.compose.uiautomator.device
-import io.getstream.chat.android.compose.uiautomator.grantPermission
-import io.getstream.chat.android.compose.uiautomator.packageName
-import io.getstream.chat.android.compose.uiautomator.testContext
 import io.getstream.chat.android.e2e.test.mockserver.MockServer
 import io.getstream.chat.android.e2e.test.robots.BackendRobot
 import io.getstream.chat.android.e2e.test.robots.ParticipantRobot
 import io.getstream.chat.android.e2e.test.rules.RetryRule
+import io.getstream.chat.android.e2e.test.uiautomator.device
+import io.getstream.chat.android.e2e.test.uiautomator.grantPermission
+import io.getstream.chat.android.e2e.test.uiautomator.packageName
+import io.getstream.chat.android.e2e.test.uiautomator.testContext
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
