+++ stream-chat-android-e2e-test/src/main/kotlin/io/getstream/chat/android/e2e/test/uiautomator/VisualOrder.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.compose.uiautomator
+package io.getstream.chat.android.e2e.test.uiautomator
 
 import androidx.test.uiautomator.BySelector
 import androidx.test.uiautomator.UiObject2
@@ -24,8 +24,10 @@ import androidx.test.uiautomator.UiObject2
  * bottom-most first. Message cells are enumerated in screen-reader reading order
  * (oldest-first), so lookups that mean "index 0 = the newest message at the visual bottom"
  * sort by on-screen position instead of relying on the enumeration order.
+ *
+ * @param timeOutMillis Maximum time to wait before returning whatever matched.
  */
-internal fun BySelector.waitToAppearBottomUp(timeOutMillis: Long = defaultTimeout): List<UiObject2> {
+public fun BySelector.waitToAppearBottomUp(timeOutMillis: Long = defaultTimeout): List<UiObject2> {
     wait(timeOutMillis)
     return device.findObjects(this).sortedByDescending { it.visibleBounds.top }
 }
@@ -33,8 +35,12 @@ internal fun BySelector.waitToAppearBottomUp(timeOutMillis: Long = defaultTimeou
 /**
  * Waits for objects matching this selector and returns the one at [withIndex], counting
  * bottom-up (index 0 = the bottom-most object on screen).
+ *
+ * @param withIndex The zero-based index counting from the bottom-most object.
+ * @param timeOutMillis Maximum time to wait before failing.
+ * @throws IllegalStateException when fewer than [withIndex] + 1 objects match.
  */
-internal fun BySelector.waitToAppearBottomUp(withIndex: Int, timeOutMillis: Long = defaultTimeout): UiObject2 {
+public fun BySelector.waitToAppearBottomUp(withIndex: Int, timeOutMillis: Long = defaultTimeout): UiObject2 {
     val objects = waitToAppearBottomUp(timeOutMillis)
     return objects.getOrNull(withIndex)
         ?: error("waitToAppearBottomUp(withIndex=$withIndex): only ${objects.size} objects matched selector: $this")
